ci: separate local cluster tests by nextest
