+++ .buildkite/scripts/build-stable.sh
@@ -21,6 +21,19 @@ partitions=$(
 EOF
 )
 
+local_cluster_partitions=$(
+  cat <<EOF
+{
+  "name": "local-cluster",
+  "command": ". ci/rust-version.sh; ci/docker-run.sh \$\$rust_stable_docker_image ci/stable/run-local-cluster-partially.sh",
+  "timeout_in_minutes": 30,
+  "agent": "$agent",
+  "parallelism": 5,
+  "retry": 3
+}
+EOF
+)
+
 localnet=$(
   cat <<EOF
 {
@@ -33,4 +46,4 @@ EOF
 )
 
 # shellcheck disable=SC2016
-group "stable" "$partitions" "$localnet"
+group "stable" "$partitions" "$local_cluster_partitions" "$localnet"
